setup (project files): group and organize files paths and imports
